+++ packages/cubejs-query-orchestrator/src/orchestrator/QueryQueue.js
@@ -869,7 +869,7 @@ export class QueryQueue {
             error: (e.stack || e).toString()
           });
           if (e instanceof TimeoutError) {
-            const queryWithCancelHandle = await queueConnection.getQueryDef(queryKeyHashed);
+            const queryWithCancelHandle = await queueConnection.getQueryDef(queryKeyHashed, queueId);
             if (queryWithCancelHandle) {
               this.logger('Cancelling query due to timeout', {
                 queueId,
@@ -887,9 +887,10 @@ export class QueryQueue {
               await this.sendCancelMessageFn(queryWithCancelHandle, queueId);
             }
           }
+        } finally {
+          // catch block can throw an exception, it's why it's important to clearInterval here
+          clearInterval(heartBeatTimer);
         }
-
-        clearInterval(heartBeatTimer);
 
         if (!(await queueConnection.setResultAndRemoveQuery(queryKeyHashed, executionResult, processingId, queueId))) {
           this.logger('Orphaned execution result', {